pyproject.toml
[project]
name = "excel-financial-agent"
version = "0.1.0"
description = "Turn PDF financial reports into investor-grade Excel dashboards instantly. This AI agent extracts key metrics, calculates EBITDA, and visualizes trends, saving you hours of manual data entry."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "llama-parse>=0.6.81",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "streamlit>=1.51.0",
    "tenacity>=9.1.2",
    "xlsxwriter>=3.2.9",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
